offline-llm-appliance/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
llama-cpp-python==0.2.20
pydantic==2.5.0
psutil==5.9.6
python-multipart==0.0.6
